+++ src/world/Server/Packets/Handlers/CalendarHandler.cpp
@@ -24,7 +24,7 @@ void WorldSession::HandleCalendarGetCalendar(WorldPacket& /*recv_data*/)
     LogDebugFlag(LF_OPCODE, "HandleCalendarGetCalendar Not handled");
 
     /* Get all events for the player */
-    uint32 guid = _player->GetGUID();
+    uint32 guid = static_cast<uint32>(_player->GetGUID());
     LogDebugFlag(LF_OPCODE, "HandleCalendarGetCalendar CMSG_CALENDAR_GET_CALENDAR for guid %u", guid);
 
 }
@@ -47,7 +47,7 @@ void WorldSession::HandleCalendarAddEvent(WorldPacket& recv_data)
     // Create an Event and save it to char db 
     LogDebugFlag(LF_OPCODE, "HandleCalendarAddEvent Not handled");
 
-    uint32 guid = _player->GetGUID();
+    uint32 guid = static_cast<uint32>(_player->GetGUID());
 
     std::string title;
     std::string description;
